requirements.txt
asgiref==3.3.4
Django==3.2.3
pytz==2021.1
sqlparse==0.4.1

djangorestframework~=3.13.1